+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Markers change color to grey when halflife is reached before deletion
 - Scale on bottom-left of the map
 - Leaflet hash handling in URL for hotlinking URLs
+- Leaflet "Locate Me" control
 
 
 ### Changed